x86info: Fix inline assembly build errors.
For the lastest git tree, there are impossible constraints for inline asm.
This patch also fixes a print format issue.

$ gcc -dumpversion
4.2.4
$ make -i
gcc -Wall -W -Wshadow -g -O2 -MMD -o cpuid.o -c cpuid.c
cpuid.c: In function 'native_cpuid':
cpuid.c:54: error: impossible register constraint in 'asm'
make: [cpuid.o] Error 1 (ignored)
scripts/testnodes
gcc -Wall -W -Wshadow -g -O2 -MMD -o lsmsr.o -c lsmsr.c
In file included from lsmsr.c:27:
msr.h: In function 'print_reg':
msr.h:126: warning: format '%16.16lx' expects type 'long unsigned int', but argument 5 has type 'uint64_t'
lsmsr.c: In function '_show_msr':
lsmsr.c:170: warning: passing argument 2 of 'get_msr_val' from incompatible pointer type
gcc -Wall -W -Wshadow -g -O2 -o lsmsr lsmsr.o cpuid.o havecpuid.o

+++ lsmsr.c
@@ -120,7 +120,7 @@ void version(void)
 	fprintf(stdout, "%s version %s\n", g.prog, VERSION);
 }
 
-int get_msr_val(unsigned int msr, unsigned long *val)
+int get_msr_val(unsigned int msr, unsigned long long *val)
 {
 	off64_t off;
 	int err;
